ix(configstore): removed unused config variables that cause static deploy to break
